(Refactoring) 유저아이디(subscribeId) 하드코딩 되어 있던 부분 변경함(prefs 에서 가져오게)

+++ app/src/main/java/com/autoever/apay_user_app/data/AppDataManager.java
@@ -234,7 +234,7 @@ public class AppDataManager implements DataManager {
     }
 
     @Override
-    public Single<BalanceResponse> doGetBalanceCall(int tokenSystemId, int subscriberId) {
+    public Single<BalanceResponse> doGetBalanceCall(int tokenSystemId, Long subscriberId) {
         return mRepoService.doGetBalanceCall(tokenSystemId, subscriberId);
     }
